Film and User Service Interfaces
