+++ src/test/java/ru/yandex/practicum/filmorate/InMemoryUserServiceTest.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class UserServiceTest {
+public class InMemoryUserServiceTest {
     @Test
     public void testFindUserById() {
         User user = new User(
